python/requirements.txt
# Connect3 Python Requirements
# Two-Tower Email Recommendation System

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Environment
python-dotenv>=1.0.0

# Database
supabase>=2.0.0

# AI/ML
openai>=1.3.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data Validation
pydantic>=2.0.0

# Templates
jinja2>=3.1.0

# Email
aiosmtplib>=2.0.0

# API
python-multipart>=0.0.6
httpx>=0.25.0
